requirements.txt
requests
pandas
pyarrow
python-dotenv
google.cloud
google-cloud-bigquery
pytz
schedule